requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-debug-toolbar==1.11
django==2.1.7
idna==2.8                 # via requests
oauthlib==3.0.1           # via requests-oauthlib, social-auth-core
pillow==5.4.1
psycopg2==2.7.7
pyjwt==1.7.1              # via social-auth-core
python3-openid==3.1.0     # via social-auth-core
pytz==2018.9              # via django
requests-oauthlib==1.2.0  # via social-auth-core
requests==2.21.0          # via requests-oauthlib, social-auth-core
six==1.12.0               # via social-auth-app-django, social-auth-core
social-auth-app-django==3.1.0
social-auth-core==3.1.0   # via social-auth-app-django
sqlparse==0.2.4           # via django-debug-toolbar
urllib3==1.24.1           # via requests
